refactor: consolidated BigQuery service classes into source and sink pipeline
