+++ extensions/common/gcp/gcp-core/src/main/java/org/eclipse/edc/gcp/bigquery/service/BigQueryFactoryImpl.java
@@ -18,16 +18,16 @@ import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.NoCredentials;
 import com.google.cloud.bigquery.BigQuery;
 import com.google.cloud.bigquery.BigQueryOptions;
-import org.eclipse.edc.gcp.common.GcpConfiguration;
+import org.eclipse.edc.gcp.bigquery.BigQueryConfiguration;
 import org.eclipse.edc.gcp.common.GcpServiceAccount;
 import org.eclipse.edc.gcp.iam.IamService;
 
 public class BigQueryFactoryImpl implements BigQueryFactory {
-    private final GcpConfiguration gcpConfiguration;
+    private final BigQueryConfiguration configuration;
     private final IamService iamService;
 
-    public BigQueryFactoryImpl(GcpConfiguration gcpConfiguration, IamService iamService) {
-        this.gcpConfiguration = gcpConfiguration;
+    public BigQueryFactoryImpl(BigQueryConfiguration configuration, IamService iamService) {
+        this.configuration = configuration;
         this.iamService = iamService;
     }
 
@@ -40,7 +40,7 @@ public class BigQueryFactoryImpl implements BigQueryFactory {
 
     private BigQuery createBigQuery(GoogleCredentials credentials) {
         var bqBuilder = BigQueryOptions.newBuilder();
-        var host = System.getProperty("EDC_GCP_BQREST");
+        var host = configuration.restEndpoint();
         if (host != null) {
             bqBuilder.setHost(host);
             bqBuilder.setLocation(host);
