+++ extensions/data-plane/data-plane-google-bigquery/src/main/java/org/eclipse/edc/connector/dataplane/gcp/bigquery/pipeline/BigQueryDataSinkFactory.java
@@ -17,10 +17,8 @@ package org.eclipse.edc.connector.dataplane.gcp.bigquery.pipeline;
 import org.eclipse.edc.connector.dataplane.gcp.bigquery.params.BigQueryRequestParamsProvider;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSink;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSinkFactory;
-import org.eclipse.edc.gcp.bigquery.service.BigQuerySinkService;
-import org.eclipse.edc.gcp.bigquery.service.BigQuerySinkServiceImpl;
+import org.eclipse.edc.gcp.bigquery.BigQueryConfiguration;
 import org.eclipse.edc.gcp.common.GcpAccessToken;
-import org.eclipse.edc.gcp.common.GcpConfiguration;
 import org.eclipse.edc.gcp.common.GcpException;
 import org.eclipse.edc.gcp.iam.IamService;
 import org.eclipse.edc.spi.monitor.Monitor;
@@ -38,24 +36,23 @@ import static org.eclipse.edc.gcp.bigquery.service.BigQueryServiceSchema.BIGQUER
  * Instantiates {@link BigQueryDataSink}
  */
 public class BigQueryDataSinkFactory implements DataSinkFactory {
-    private final GcpConfiguration gcpConfiguration;
+    private final BigQueryConfiguration configuration;
     private final ExecutorService executorService;
     private final Monitor monitor;
     private final Vault vault;
     private final TypeManager typeManager;
     private final BigQueryRequestParamsProvider requestParamsProvider;
-    private BigQuerySinkService bqSinkService;
     private IamService iamService;
 
     public BigQueryDataSinkFactory(
-                               GcpConfiguration gcpConfiguration,
-                               ExecutorService executorService,
-                               Monitor monitor,
-                               Vault vault,
-                               TypeManager typeManager,
-                               BigQueryRequestParamsProvider requestParamsProvider,
-                               IamService iamService) {
-        this.gcpConfiguration = gcpConfiguration;
+            BigQueryConfiguration configuration,
+            ExecutorService executorService,
+            Monitor monitor,
+            Vault vault,
+            TypeManager typeManager,
+            BigQueryRequestParamsProvider requestParamsProvider,
+            IamService iamService) {
+        this.configuration = configuration;
         this.executorService = executorService;
         this.monitor = monitor;
         this.vault = vault;
@@ -92,18 +89,7 @@ public class BigQueryDataSinkFactory implements DataSinkFactory {
         monitor.info("BigQuery Data Sink Factory " + message.getId());
         var params = requestParamsProvider.provideSinkParams(message);
         var target = params.getTarget();
-
-        if (bqSinkService != null) {
-            monitor.info("BigQuery Data Sink Factory reusing existing connecting service");
-            return BigQueryDataSink.Builder.newInstance()
-                    .requestId(message.getId())
-                    .executorService(executorService)
-                    .monitor(monitor)
-                    .sinkService(bqSinkService)
-                    .build();
-        }
-
-        var bqSinkServiceBuilder = BigQuerySinkServiceImpl.Builder.newInstance(gcpConfiguration, target, monitor);
+        var dataSinkBuilder = BigQueryDataSink.Builder.newInstance();
         var keyName = message.getDestinationDataAddress().getKeyName();
         var usingProvisionerAccessToken = false;
         if (keyName != null && !keyName.isEmpty()) {
@@ -112,7 +98,7 @@ public class BigQueryDataSinkFactory implements DataSinkFactory {
                 var gcsAccessToken = typeManager.readValue(credentialsContent,
                         GcpAccessToken.class);
                 var credentials  = iamService.getCredentials(gcsAccessToken);
-                bqSinkServiceBuilder.credentials(credentials);
+                dataSinkBuilder.credentials(credentials);
                 usingProvisionerAccessToken = true;
                 monitor.info("BigQuery Data Sink using provisioner's access token");
             } else {
@@ -120,54 +106,19 @@ public class BigQueryDataSinkFactory implements DataSinkFactory {
             }
         }
 
-        if (!usingProvisionerAccessToken && System.getProperty("EDC_GCP_BQRPC") == null) {
+        if (!usingProvisionerAccessToken && configuration.rpcEndpoint() == null) {
             var serviceAccount = iamService.getServiceAccount(params.getServiceAccountName());
             var credentials = iamService.getCredentials(serviceAccount, IamService.BQ_SCOPE);
-            bqSinkServiceBuilder.credentials(credentials);
+            dataSinkBuilder.credentials(credentials);
             monitor.info("BigQuery Data Sink using service account from data address: " + serviceAccount.getName());
         }
 
-        var newBqSinkService = bqSinkServiceBuilder.build();
-        monitor.info("BigQuery Data Sink Factory connecting service ready");
-
-        return BigQueryDataSink.Builder.newInstance()
+        return dataSinkBuilder
             .requestId(message.getId())
             .executorService(executorService)
             .monitor(monitor)
-            .sinkService(newBqSinkService)
+            .bigQueryTarget(target)
+            .configuration(configuration)
             .build();
-    }
-
-    public static class Builder {
-        private BigQueryDataSinkFactory factory;
-
-        private Builder(
-                GcpConfiguration gcpConfiguration,
-                ExecutorService executorService,
-                Monitor monitor,
-                Vault vault,
-                TypeManager typeManager,
-                BigQueryRequestParamsProvider requestParamsProvider, IamService iamService) {
-            factory = new BigQueryDataSinkFactory(gcpConfiguration, executorService, monitor, vault, typeManager, requestParamsProvider, iamService);
-        }
-
-        public static Builder newInstance(
-                GcpConfiguration gcpConfiguration,
-                ExecutorService executorService,
-                Monitor monitor,
-                Vault vault,
-                TypeManager typeManager,
-                BigQueryRequestParamsProvider requestParamsProvider, IamService iamService) {
-            return new Builder(gcpConfiguration, executorService, monitor, vault, typeManager, requestParamsProvider, iamService);
-        }
-
-        public Builder sinkService(BigQuerySinkService bqSinkService) {
-            factory.bqSinkService = bqSinkService;
-            return this;
-        }
-
-        public BigQueryDataSinkFactory build() {
-            return factory;
-        }
     }
 }
